requirements.txt
gTTS==2.5.4
pygame==2.6.1
pynput==1.8.1
python-dotenv==1.1.1
pyttsx3==2.98
speechrecognition==3.14.3
spotipy==2.25.1
